rebase new tests properly, hopefully will work now
