+++ test/python/golden/test_ttir_ops.py
@@ -30,27 +30,57 @@ def abs(in0: Operand, builder: TTIRBuilder, unit_attrs: Optional[List[str]] = No
     return builder.abs(in0, unit_attrs=unit_attrs)
 
 
+def logical_not(
+    in0: Operand,
+    builder: TTIRBuilder,
+    shape: Shape,
+    dtype: torch.dtype,
+    unit_attrs: Optional[List[str]] = None,
+):
+    randn_tensor = torch.randn(shape, dtype=torch.float32)
+    input_tensor = randn_tensor.uniform_(-10.0, 10.0)
+    input_tensor[torch.abs(input_tensor) < 4.0] = 0.0
+    input_tensor = input_tensor.to(dtype)
+    # Torch returns bool tensor but ttnn doesn't have bool type, convert to input dtype
+    golden_output_tensor = torch.logical_not(input_tensor).to(dtype)
+    builder.set_graph_input_output(
+        [input_tensor], [golden_output_tensor], override=True
+    )
+    return builder.logical_not(in0, unit_attrs=unit_attrs)
+
+
 # TODO (wenbinlyuTT): test int32 once untilize issue is fixed
 @pytest.mark.parametrize("shape", [(128, 128)], ids=shape_str)
 @pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
 @pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
 def test_logical_not(shape: Shape, dtype: torch.dtype, target: str, request):
-    def logical_not(
+    def logical_not_wrapper(
         in0: Operand, builder: TTIRBuilder, unit_attrs: Optional[List[str]] = None
     ):
-        randn_tensor = torch.randn(shape, dtype=torch.float32)
-        input_tensor = randn_tensor.uniform_(-10.0, 10.0)
-        input_tensor[torch.abs(input_tensor) < 4.0] = 0.0
-        input_tensor = input_tensor.to(dtype)
-        # Torch returns bool tensor but ttnn doesn't have bool type, convert to input dtype
-        golden_output_tensor = torch.logical_not(input_tensor).to(dtype)
-        builder.set_graph_input_output(
-            [input_tensor], [golden_output_tensor], override=True
-        )
-        return builder.logical_not(in0, unit_attrs=unit_attrs)
+        return logical_not(in0, builder, shape, dtype, unit_attrs)
 
     compile_to_flatbuffer(
-        logical_not,
+        logical_not_wrapper,
+        [shape],
+        [dtype],
+        test_base=request.node.name,
+        output_root=request.config.getoption("--path"),
+        system_desc_path=request.config.getoption("--sys-desc"),
+        target=target,
+    )
+
+
+@pytest.mark.parametrize("shape", [(128, 128)], ids=shape_str)
+@pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
+@pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
+def test_hoisted_logical_not(shape: Shape, dtype: torch.dtype, target: str, request):
+    def hoisted_logical_not_wrapper(
+        in0: Operand, builder: TTIRBuilder, unit_attrs: Optional[List[str]] = None
+    ):
+        return logical_not(in0, builder, shape, dtype, unit_attrs=["ttir.should_hoist"])
+
+    compile_to_flatbuffer(
+        hoisted_logical_not_wrapper,
         [shape],
         [dtype],
         test_base=request.node.name,
@@ -482,29 +512,63 @@ def lt(
     return builder.lt(in0, in1, unit_attrs=unit_attrs)
 
 
+def div(
+    in0: Operand,
+    in1: Operand,
+    builder: TTIRBuilder,
+    shape: Shape,
+    dtype: torch.dtype,
+    unit_attrs: Optional[List[str]] = None,
+):
+    dividend_tensor = torch.randn(shape, dtype=dtype)
+    divisor_tensor = torch.randn(shape, dtype=dtype)
+    dividend_tensor[torch.abs(dividend_tensor) < 0.01] = 0.03
+    divisor_tensor[torch.abs(divisor_tensor) < 0.01] = -0.03
+    output_golden = torch.div(dividend_tensor, divisor_tensor)
+    builder.set_graph_input_output(
+        [dividend_tensor, divisor_tensor], [output_golden], override=True
+    )
+    return builder.div(in0, in1, unit_attrs=unit_attrs)
+
+
 # TODO (wenbinlyuTT): fix f32 accuracy issue for small values
 @pytest.mark.parametrize("shape", [(128, 128)], ids=shape_str)
 @pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
 @pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
 def test_div(shape: Shape, dtype: torch.dtype, target: str, request):
-    def div(
+    def div_wrapper(
         in0: Operand,
         in1: Operand,
         builder: TTIRBuilder,
         unit_attrs: Optional[List[str]] = None,
     ):
-        dividend_tensor = torch.randn(shape, dtype=dtype)
-        divisor_tensor = torch.randn(shape, dtype=dtype)
-        dividend_tensor[torch.abs(dividend_tensor) < 0.01] = 0.03
-        divisor_tensor[torch.abs(divisor_tensor) < 0.01] = -0.03
-        output_golden = torch.div(dividend_tensor, divisor_tensor)
-        builder.set_graph_input_output(
-            [dividend_tensor, divisor_tensor], [output_golden], override=True
-        )
-        return builder.div(in0, in1, unit_attrs=unit_attrs)
+        return div(in0, in1, builder, shape, dtype, unit_attrs)
 
     compile_to_flatbuffer(
-        div,
+        div_wrapper,
+        [shape, shape],
+        [dtype, dtype],
+        test_base=request.node.name,
+        output_root=request.config.getoption("--path"),
+        system_desc_path=request.config.getoption("--sys-desc"),
+        target=target,
+    )
+
+
+@pytest.mark.parametrize("shape", [(128, 128)], ids=shape_str)
+@pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
+@pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
+def test_hoisted_div(shape: Shape, dtype: torch.dtype, target: str, request):
+    def div_wrapper(
+        in0: Operand,
+        in1: Operand,
+        builder: TTIRBuilder,
+        unit_attrs: Optional[List[str]] = None,
+    ):
+        return div(in0, in1, builder, shape, dtype, unit_attrs=["ttir.should_hoist"])
+
+    compile_to_flatbuffer(
+        div_wrapper,
         [shape, shape],
         [dtype, dtype],
         test_base=request.node.name,
@@ -541,6 +605,25 @@ def minimum(
     return builder.minimum(in0, in1, unit_attrs=unit_attrs)
 
 
+def pow(
+    in0: Operand,
+    in1: Operand,
+    builder: TTIRBuilder,
+    shape: Shape,
+    dtype: torch.dtype,
+    unit_attrs: Optional[List[str]] = None,
+):
+    randn_base_tensor = torch.randn(shape, dtype=dtype)
+    randn_exponent_tensor = torch.randn(shape, dtype=dtype)
+    if torch.is_floating_point(randn_exponent_tensor):
+        randn_base_tensor = torch.abs(randn_base_tensor)
+    output_golden = torch.pow(randn_base_tensor, randn_exponent_tensor)
+    builder.set_graph_input_output(
+        [randn_base_tensor, randn_exponent_tensor], [output_golden], override=True
+    )
+    return builder.pow(in0, in1, unit_attrs=unit_attrs)
+
+
 @pytest.mark.parametrize("shapes", [[(10, 64, 32), (32, 128), (128,)]])
 def test_linear(shapes: List[Shape], request):
     def linear(
@@ -566,24 +649,40 @@ def test_linear(shapes: List[Shape], request):
 @pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
 @pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
 def test_pow(shape: Shape, dtype: torch.dtype, target: str, request):
-    def pow(
+    def pow_wrapper(
         in0: Operand,
         in1: Operand,
         builder: TTIRBuilder,
         unit_attrs: Optional[List[str]] = None,
     ):
-        randn_base_tensor = torch.randn(shape, dtype=dtype)
-        randn_exponent_tensor = torch.randn(shape, dtype=dtype)
-        if torch.is_floating_point(randn_exponent_tensor):
-            randn_base_tensor = torch.abs(randn_base_tensor)
-        output_golden = torch.pow(randn_base_tensor, randn_exponent_tensor)
-        builder.set_graph_input_output(
-            [randn_base_tensor, randn_exponent_tensor], [output_golden], override=True
-        )
-        return builder.pow(in0, in1, unit_attrs=unit_attrs)
+        return pow(in0, in1, builder, shape, dtype, unit_attrs)
 
     compile_to_flatbuffer(
-        pow,
+        pow_wrapper,
+        [shape, shape],
+        [dtype, dtype],
+        test_base=request.node.name,
+        output_root=request.config.getoption("--path"),
+        system_desc_path=request.config.getoption("--sys-desc"),
+        target=target,
+    )
+
+
+@pytest.mark.fails_golden
+@pytest.mark.parametrize("shape", [(128, 128)], ids=shape_str)
+@pytest.mark.parametrize("dtype", [torch.float32], ids=["f32"])
+@pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
+def test_hoisted_pow(shape: Shape, dtype: torch.dtype, target: str, request):
+    def hoisted_pow_wrapper(
+        in0: Operand,
+        in1: Operand,
+        builder: TTIRBuilder,
+        unit_attrs: Optional[List[str]] = None,
+    ):
+        return pow(in0, in1, builder, shape, dtype, unit_attrs=["ttir.should_hoist"])
+
+    compile_to_flatbuffer(
+        hoisted_pow_wrapper,
         [shape, shape],
         [dtype, dtype],
         test_base=request.node.name,
@@ -1205,6 +1304,7 @@ def reduce_or(
         in0, dim_args=dim_args, keep_dim=keep_dim, unit_attrs=unit_attrs
     )
 
+
 @pytest.mark.skip(
     "Generated flatbuffer will currently fail to run due to only floats being supported by the runtime. See issue #1775"
 )
@@ -1614,7 +1714,9 @@ def create_hoisted_where_op(op_func, name):
     """Create a hoisted version of the where operation"""
 
     def hoisted_op(condition, x, y, builder, **kwargs):
-        return op_func(condition, x, y, builder, unit_attrs=["should_hoist"], **kwargs)
+        return op_func(
+            condition, x, y, builder, unit_attrs=["ttir.should_hoist"], **kwargs
+        )
 
     hoisted_op.__name__ = f"hoisted_{name}"
     return hoisted_op
@@ -1630,7 +1732,13 @@ def create_hoisted_slice_op(op_func, name):
         ends = DenseI32ArrayAttr.get([10, 10])
         steps = DenseI32ArrayAttr.get([1, 1])
         return op_func(
-            in0, begins, ends, steps, builder, unit_attrs=["should_hoist"], **kwargs
+            in0,
+            begins,
+            ends,
+            steps,
+            builder,
+            unit_attrs=["ttir.should_hoist"],
+            **kwargs,
         )
 
     hoisted_op.__name__ = f"hoisted_{name}"
@@ -1648,7 +1756,7 @@ def create_hoisted_reduce_op(op_func, name):
             in0,
             builder,
             dim_args=default_dim_args,
-            unit_attrs=["should_hoist"],
+            unit_attrs=["ttir.should_hoist"],
             **kwargs,
         )
 
@@ -1668,7 +1776,6 @@ hoisted_unary_ops = [
     create_hoisted_unary_op(sigmoid, "sigmoid"),
     create_hoisted_unary_op(sin, "sin"),
     create_hoisted_unary_op(cos, "cos"),
-    create_hoisted_unary_op(logical_not, "logical_not"),
     pytest.param(
         create_hoisted_unary_op(max, "max"),
         marks=pytest.mark.skip(
@@ -1682,12 +1789,6 @@ hoisted_unary_ops = [
             reason="Softmax does not lower to loops properly https://github.com/tenstorrent/tt-mlir/issues/3232"
         ),
     ),
-    pytest.param(
-        create_hoisted_unary_op(max, "max"),
-        marks=pytest.mark.skip(
-            reason="max and torch max do not align, https://github.com/tenstorrent/tt-mlir/issues/3850"
-        ),
-    ),
     create_hoisted_unary_op(reshape, "reshape"),
     create_hoisted_unary_op(transpose, "transpose"),
 ]
@@ -1697,8 +1798,6 @@ hoisted_binary_ops = [
     create_hoisted_binary_op(add, "add"),
     create_hoisted_binary_op(multiply, "multiply"),
     create_hoisted_binary_op(subtract, "subtract"),
-    create_hoisted_binary_op(div, "div"),
-    create_hoisted_binary_op(pow, "pow"),
     create_hoisted_binary_op(eq, "equal"),
     create_hoisted_binary_op(ne, "not_equal"),
     create_hoisted_binary_op(gt, "greater_than"),
@@ -1819,7 +1918,7 @@ def test_hoisted_slice(
         in0: Operand, builder: TTIRBuilder, unit_attrs: Optional[List[str]] = None
     ):
         # Now use the slice operation with the CPU hoisting attribute
-        return builder.slice(in0, begins, ends, step, unit_attrs=["should_hoist"])
+        return builder.slice(in0, begins, ends, step, unit_attrs=["ttir.should_hoist"])
 
     compile_to_flatbuffer(
         slice_wrapper,
@@ -1836,7 +1935,7 @@ def test_hoisted_slice(
 @pytest.mark.parametrize("target", ["ttnn", "ttmetal"])
 def test_hoisted_where(shapes, request, target: str):
     def where_wrapper(condition: Operand, x: Operand, y: Operand, builder: TTIRBuilder):
-        return builder.where(condition, x, y, unit_attrs=["should_hoist"])
+        return builder.where(condition, x, y, unit_attrs=["ttir.should_hoist"])
 
     where_wrapper.__name__ = "hoisted_where"
 
@@ -1864,7 +1963,7 @@ def test_hoisted_reshape(shapes, request, target: str):
     input_shape, output_shape = shapes
 
     def reshape_wrapper(in0: Operand, builder: TTIRBuilder):
-        return builder.reshape(in0, output_shape, unit_attrs=["should_hoist"])
+        return builder.reshape(in0, output_shape, unit_attrs=["ttir.should_hoist"])
 
     reshape_wrapper.__name__ = "hoisted_reshape"
 
@@ -1896,7 +1995,9 @@ def test_hoisted_transpose(shapes_and_dims, request, target: str):
         # For 3D tensors with permutation [2, 1, 0], swap dimensions 0 and 2
         dim0 = dims[0]
         dim1 = dims[1]
-        return builder.transpose(in0, dim0=dim0, dim1=dim1, unit_attrs=["should_hoist"])
+        return builder.transpose(
+            in0, dim0=dim0, dim1=dim1, unit_attrs=["ttir.should_hoist"]
+        )
 
     transpose_wrapper.__name__ = "hoisted_transpose"
 
@@ -2097,7 +2198,7 @@ def test_hoisted_reduce_or(shape: Shape, dim_args: List[int], target: str, reque
         in0: Operand, builder: TTIRBuilder, unit_attrs: Optional[List[str]] = None
     ):
         return reduce_or(
-            in0, builder, dim_args, keep_dim=True, unit_attrs=["should_hoist"]
+            in0, builder, dim_args, keep_dim=True, unit_attrs=["ttir.should_hoist"]
         )
 
     compile_to_flatbuffer(
@@ -2132,7 +2233,7 @@ def test_hoisted_broadcast(shapes_and_broadcast_dims, request, target: str):
         unit_attrs: Optional[List[str]] = None,
     ):
         return broadcast(
-            in0, in1, builder, broadcast_dimensions, unit_attrs=["should_hoist"]
+            in0, in1, builder, broadcast_dimensions, unit_attrs=["ttir.should_hoist"]
         )
 
     broadcast_wrapper.__name__ = "hoisted_broadcast"
@@ -2268,7 +2369,7 @@ def test_hoisted_gather(
             start_index_map,
             offset_dims,
             slice_sizes,
-            unit_attrs=["should_hoist"],
+            unit_attrs=["ttir.should_hoist"],
         )
 
     compile_to_flatbuffer(
@@ -2321,7 +2422,7 @@ def test_hoisted_dot_general(
             contract_dims_lhs,
             batch_dims_rhs,
             contract_dims_rhs,
-            unit_attrs=["should_hoist"],
+            unit_attrs=["ttir.should_hoist"],
         )
 
     compile_to_flatbuffer(
